+++ sklearn/preprocessing/label.py
@@ -2,9 +2,15 @@
 #          Mathieu Blondel <mathieu@mblondel.org>
 #          Olivier Grisel <olivier.grisel@ensta.org>
 #          Andreas Mueller <amueller@ais.uni-bonn.de>
+#          Joel Nothman <joel.nothman@gmail.com>
 # License: BSD 3 clause
 
+from collections import defaultdict
+import itertools
+import array
+
 import numpy as np
+import scipy.sparse as sp
 
 from ..base import BaseEstimator, TransformerMixin
 
@@ -205,11 +211,11 @@ class LabelBinarizer(BaseEstimator, TransformerMixin):
     array([[1, 0, 0, 0],
            [0, 0, 0, 1]])
 
-    >>> lb.fit_transform([(1, 2), (3,)])
-    array([[1, 1, 0],
-           [0, 0, 1]])
+    >>> import numpy as np
+    >>> lb.fit(np.array([[0, 1, 1], [1, 0, 0]]))
+    LabelBinarizer(neg_label=0, pos_label=1)
     >>> lb.classes_
-    array([1, 2, 3])
+    array([0, 1, 2])
     >>> lb.multilabel_
     True
 
@@ -393,12 +399,6 @@ def label_binarize(y, classes, multilabel=False, neg_label=0, pos_label=1):
     array([[1, 0, 0, 0],
            [0, 1, 0, 0]])
 
-    >>> label_binarize([(1, 2), (6,), ()], multilabel=True,
-    ...                classes=[1, 6, 4, 2])
-    array([[1, 0, 0, 1],
-           [0, 1, 0, 0],
-           [0, 0, 0, 0]])
-
     See also
     --------
     label_binarize : function to perform the transform operation of
@@ -445,3 +445,136 @@ def label_binarize(y, classes, multilabel=False, neg_label=0, pos_label=1):
         else:
             # Only one class, returns a matrix with all negative labels.
             return Y
+
+
+class MultiLabelBinarizer(BaseEstimator, TransformerMixin):
+    """Transform between iterable of iterables and a multilabel format
+
+    Although a list of sets or tuples is a very intuitive format for multilabel
+    data, it is unwieldy to process. This transformer converts between this
+    intuitive format and the supported multilabel format: a (samples x classes)
+    binary matrix indicating the presence of a class label.
+
+    Parameters
+    ----------
+    classes : array-like of shape [n_classes] (optional)
+        Indicates an ordering for the class labels
+
+    Attributes
+    ----------
+    classes_ : array of labels
+        A copy of the `classes` parameter where provided, or otherwise, the
+        sorted set of classes found when fitting.
+
+    >>> lsb = MultiLabelBinarizer()
+    >>> lsb.fit_transform([(1, 2), (3,)])
+    array([[1, 1, 0],
+           [0, 0, 1]])
+    >>> lsb.classes_          # doctest: +ELLIPSIS
+    array([1, 2, 3]...)
+    >>> lsb.fit_transform([{1, 2}, {3,}])
+    array([[1, 1, 0],
+           [0, 0, 1]])
+    """
+    def __init__(self, classes=None):
+        self.classes = classes
+
+    def fit(self, y):
+        """Fit the label sets binarizer, storing classes_
+
+        Parameters
+        ----------
+        y : iterable of iterables
+            A set of labels (any orderable and hashable object) for each
+            sample. If the `classes` parameter is set, `y` will not be
+            iterated.
+
+        Returns
+        -------
+        self : returns this MultiLabelBinarizer instance
+        """
+        if self.classes is None:
+            classes = sorted(set(itertools.chain.from_iterable(y)))
+        else:
+            classes = self.classes
+        self.classes_ = np.empty(len(classes), dtype=object)
+        self.classes_[:] = classes
+        return self
+
+    def fit_transform(self, y):
+        """Fit the label sets binarizer and transform the given label sets
+
+        Parameters
+        ----------
+        y : iterable of iterables
+            A set of labels (any orderable and hashable object) for each
+            sample. If the `classes` parameter is set, `y` will not be
+            iterated.
+
+        Returns
+        -------
+        y_indicator : matrix of shape (n_samples, n_classes)
+            A matrix such that `y_indicator[i, j] = 1` iff `classes_[j]` is in
+            `y[i]`, and 0 otherwise.
+        """
+        if self.classes is not None:
+            return self.fit(y).transform(y)
+
+        # Automatically increment on new class
+        class_mapping = defaultdict(int)
+        class_mapping.default_factory = class_mapping.__len__
+        yt = self._transform(y, class_mapping)
+
+        # sort classes and reorder columns
+        tmp = sorted(class_mapping, key=class_mapping.get)
+        # (make safe for tuples)
+        class_mapping = np.empty(len(tmp), dtype=object)
+        class_mapping[:] = tmp
+        self.classes_, inverse = np.unique(class_mapping, return_inverse=True)
+        yt.indices = np.take(inverse, yt.indices)
+        return yt.toarray()
+
+    def transform(self, y):
+        """Transform the given label sets
+
+        Parameters
+        ----------
+        y : iterable of iterables
+            A set of labels (any orderable and hashable object) for each
+            sample. If the `classes` parameter is set, `y` will not be
+            iterated.
+
+        Returns
+        -------
+        y_indicator : matrix of shape (n_samples, n_classes)
+            A matrix such that `y_indicator[i, j] = 1` iff `classes_[j]` is in
+            `y[i]`, and 0 otherwise.
+        """
+        yt = self._transform(y, {k: i for i, k in enumerate(self.classes_)})
+        return yt.toarray()
+
+    def _transform(self, y, class_mapping):
+        indices = array.array('i')
+        indptr = array.array('i', [0])
+        for labels in y:
+            indices.extend(class_mapping[label] for label in labels)
+            indptr.append(len(indices))
+        data = np.ones(len(indices), dtype=int)
+        return sp.csr_matrix((data, indices, indptr),
+                             shape=(len(indptr) - 1, len(class_mapping)))
+
+    def inverse_transform(self, yt):
+        """Transform the given indicator matrix into label sets
+
+        Parameters
+        ----------
+        yt : matrix of shape (n_samples, n_classes)
+            A matrix containing only 1s ands 0s.
+
+        Returns
+        -------
+        y : list of tuples
+            The set of labels for each sample such that `y[i]` consists of
+            `classes_[j]` for each `yt[i, j] == 1`.
+        """
+        return [tuple(self.classes_.compress(indicators)) for indicators in yt]